Improvements to the fonoster client hooks
