+++ mods/webui/src/common/sdk/hooks/useFonosterClient.tsx
@@ -45,7 +45,7 @@ export function useFonosterClient(): FonosterClient {
   const { client, isInitialized, session, authClient } = useContext(FonosterContext);
   const { notifyError } = useNotification();
   const [hasCheckedSession, setHasCheckedSession] = useState(false);
-  
+
   useEffect(() => {
     let isMounted = true;
 
@@ -76,14 +76,12 @@ export function useFonosterClient(): FonosterClient {
       if (!authClient || !client) {
         return undefined;
       }
-      return await authClient.executeWithRefresh(() => {
-        return client.verifyCode(params as any);
-      });
+      return await authClient.executeWithRefresh(() => client.verifyCode(params as any));
     } catch (error: any) {
       notifyError(error as ErrorType);
       return undefined;
     }
-  }, [authClient, client, notifyError]);
+  }, [authClient, client]);
 
   /**
    * Sends a verification code
@@ -91,16 +89,15 @@ export function useFonosterClient(): FonosterClient {
   const sendVerificationCode = useCallback(async (params: SendVerificationCode): Promise<any> => {
     try {
       if (!authClient || !client) {
+        console.error('authClient or client is not initialized');
         return undefined;
       }
-      return await authClient.executeWithRefresh(() => {
-        return client.sendVerificationCode(params as any);
-      });
+      return await authClient.executeWithRefresh(() => client.sendVerificationCode(params as any));
     } catch (error: any) {
       notifyError(error as ErrorType);
       return undefined;
     }
-  }, [authClient, client, notifyError]);
+  }, [client]);
 
   /**
    * Sets the access key ID
